requirements.txt
cityscapesscripts
cycler
gdown
humanfriendly
kiwisolver
kornia
matplotlib
numpy
opencv-python
pandas
Pillow
prettytable
pyparsing
pytz
PyYAML
scipy
seaborn
timm
torch==1.9.1+cu102
torchvision==0.10.2+cu102
tqdm
typing-extensions
wcwidth
yapf
